Add one test
