+++ legend-engine-xt-persistence-component/legend-engine-xt-persistence-component-relational-h2/src/test/java/org/finos/legend/engine/persistence/components/ingestmode/nontemporal/AppendOnlyTest.java
@@ -266,106 +266,6 @@ class AppendOnlyTest extends BaseTest
     }
 
     /*
-    Scenario: Test Append Only when Staging table has extra column which should be handled invoke schema evolution
-    */
-    @Test
-    void testAppendOnlyWithAddColumnEvolution() throws Exception
-    {
-        DatasetDefinition mainTable = TestUtils.getSchemaEvolMainTableWithMissingColumn();
-        DatasetDefinition stagingTable = TestUtils.getBasicStagingTable();
-
-        // Create staging table
-        createStagingTable(stagingTable);
-
-        // Generate the milestoning object
-        AppendOnly ingestMode = AppendOnly.builder()
-            .digestField(digestName)
-            .deduplicationStrategy(FilterDuplicates.builder().build())
-            .auditing(NoAuditing.builder().build())
-            .build();
-
-        PlannerOptions options = PlannerOptions.builder().cleanupStagingData(false).collectStatistics(true).enableSchemaEvolution(true).build();
-        Datasets datasets = Datasets.of(mainTable, stagingTable);
-
-        String[] schema = new String[]{idName, nameName, incomeName, startTimeName, expiryDateName, digestName};
-
-        // ------------ Perform incremental (append) milestoning Pass1 ------------------------
-        String dataPass1 = basePath + "input/vanilla_case/data_pass1.csv";
-        String expectedDataPass1 = basePath + "expected/vanilla_case/expected_pass1.csv";
-        // 1. Load staging table
-        loadBasicStagingData(dataPass1);
-        // 2. Execute plans and verify results
-        Map<String, Object> expectedStats = new HashMap<>();
-        expectedStats.put(StatisticName.INCOMING_RECORD_COUNT.name(), 3);
-        expectedStats.put(StatisticName.ROWS_DELETED.name(), 0);
-        expectedStats.put(StatisticName.ROWS_UPDATED.name(), 0);
-        expectedStats.put(StatisticName.ROWS_TERMINATED.name(), 0);
-        IngestorResult result = executePlansAndVerifyResults(ingestMode, options, datasets, schema, expectedDataPass1, expectedStats);
-        // 3. Assert that the staging table is NOT truncated
-        List<Map<String, Object>> stagingTableList = h2Sink.executeQuery("select * from \"TEST\".\"staging\"");
-        Assertions.assertEquals(stagingTableList.size(), 3);
-
-        // ------------ Perform incremental (append) milestoning Pass2 ------------------------
-        String dataPass2 = basePath + "input/vanilla_case/data_pass2.csv";
-        String expectedDataPass2 = basePath + "expected/vanilla_case/expected_pass2.csv";
-        // 1. Load staging table
-        loadBasicStagingData(dataPass2);
-        // 2. Execute plans and verify results
-        executePlansAndVerifyResults(ingestMode, options, result.updatedDatasets(), schema, expectedDataPass2, expectedStats);
-    }
-
-    /*
-    Scenario: Test AppendOnly with implicit data type change schema evolution
-    */
-    @Test
-    void testAppendOnlyWithImplicitDataTypeChange() throws Exception
-    {
-        DatasetDefinition mainTable = TestUtils.getBasicMainTable();
-        DatasetDefinition stagingTable = TestUtils.getStagingTableForImplicitSchemaEvolution();
-
-        // Create staging table
-        createStagingTable(stagingTable);
-
-        // Generate the milestoning object
-        AppendOnly ingestMode = AppendOnly.builder()
-            .digestField(digestName)
-            .deduplicationStrategy(FilterDuplicates.builder().build())
-            .auditing(NoAuditing.builder().build())
-            .build();
-
-        PlannerOptions options = PlannerOptions.builder().cleanupStagingData(false).collectStatistics(true).enableSchemaEvolution(true).build();
-        Datasets datasets = Datasets.of(mainTable, stagingTable);
-
-        String[] schema = new String[]{idName, nameName, incomeName, startTimeName, expiryDateName, digestName};
-
-        // ------------ Perform incremental (append) milestoning Pass1 ------------------------
-        String dataPass1 = basePath + "input/vanilla_case/data_pass1.csv";
-        String expectedDataPass1 = basePath + "expected/vanilla_case/expected_pass1.csv";
-        // 1. Load staging table
-        loadBasicStagingData(dataPass1);
-        // 2. Execute plans and verify results
-        Map<String, Object> expectedStats = new HashMap<>();
-        expectedStats.put(StatisticName.INCOMING_RECORD_COUNT.name(), 3);
-        expectedStats.put(StatisticName.ROWS_DELETED.name(), 0);
-        expectedStats.put(StatisticName.ROWS_UPDATED.name(), 0);
-        expectedStats.put(StatisticName.ROWS_TERMINATED.name(), 0);
-
-        executePlansAndVerifyResults(ingestMode, options, datasets, schema, expectedDataPass1, expectedStats);
-        // 3. Assert that the staging table is NOT truncated
-        List<Map<String, Object>> stagingTableList = h2Sink.executeQuery("select * from \"TEST\".\"staging\"");
-        Assertions.assertEquals(stagingTableList.size(), 3);
-
-        // ------------ Perform incremental (append) milestoning Pass2 ------------------------
-        String dataPass2 = basePath + "input/vanilla_case/data_pass2.csv";
-        String expectedDataPass2 = basePath + "expected/vanilla_case/expected_pass2.csv";
-        // 1. Load staging table
-        loadBasicStagingData(dataPass2);
-        // 2. Execute plans and verify results
-        executePlansAndVerifyResults(ingestMode, options, datasets, schema, expectedDataPass2, expectedStats);
-    }
-
-
-    /*
     Scenario: Test AppendOnly with Filter Duplicates and Data Splits enabled
     */
     @Test
