+++ legend-engine-xt-persistence-component/legend-engine-xt-persistence-component-relational-h2/src/test/java/org/finos/legend/engine/persistence/components/logicalplan/operations/SchemaEvolutionTest.java
@@ -1,0 +1,132 @@
+// Copyright 2022 Goldman Sachs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.finos.legend.engine.persistence.components.logicalplan.operations;
+
+import org.finos.legend.engine.persistence.components.BaseTest;
+import org.finos.legend.engine.persistence.components.TestUtils;
+import org.finos.legend.engine.persistence.components.common.Datasets;
+import org.finos.legend.engine.persistence.components.common.StatisticName;
+import org.finos.legend.engine.persistence.components.ingestmode.AppendOnly;
+import org.finos.legend.engine.persistence.components.ingestmode.audit.NoAuditing;
+import org.finos.legend.engine.persistence.components.ingestmode.deduplication.FilterDuplicates;
+import org.finos.legend.engine.persistence.components.logicalplan.datasets.DatasetDefinition;
+import org.finos.legend.engine.persistence.components.planner.PlannerOptions;
+import org.junit.jupiter.api.Test;
+
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import static org.finos.legend.engine.persistence.components.TestUtils.assertTableColumnsEquals;
+import static org.finos.legend.engine.persistence.components.TestUtils.digestName;
+import static org.finos.legend.engine.persistence.components.TestUtils.expiryDateName;
+import static org.finos.legend.engine.persistence.components.TestUtils.idName;
+import static org.finos.legend.engine.persistence.components.TestUtils.incomeName;
+import static org.finos.legend.engine.persistence.components.TestUtils.nameName;
+import static org.finos.legend.engine.persistence.components.TestUtils.startTimeName;
+
+class SchemaEvolutionTest extends BaseTest
+{
+    private String basePathForInput = "src/test/resources/data/schema-evolution/input/";
+    private String basePathForExpected = "src/test/resources/data/schema-evolution/expected/";
+
+    @Test
+    void testAddColumn() throws Exception
+    {
+        DatasetDefinition mainTable = TestUtils.getSchemaEvolutionBeforeAddMainTable();
+        DatasetDefinition stagingTable = TestUtils.getSchemaEvolutionBeforeAddStagingTable();
+
+        String[] schema = new String[]{idName, nameName, incomeName, startTimeName, expiryDateName, digestName};
+
+        // Create staging table
+        createStagingTable(stagingTable);
+
+        AppendOnly ingestMode = AppendOnly.builder()
+            .digestField(digestName)
+            .deduplicationStrategy(FilterDuplicates.builder().build())
+            .auditing(NoAuditing.builder().build())
+            .build();
+
+        PlannerOptions options = PlannerOptions.builder().cleanupStagingData(false).collectStatistics(true).enableSchemaEvolution(true).build();
+        Datasets datasets = Datasets.of(mainTable, stagingTable);
+
+        // ------------ Perform Pass1 ------------------------
+        String dataPass1 = basePathForInput + "add_data_pass.csv";
+        String expectedDataPass1 = basePathForExpected + "add_expected_pass.csv";
+        // 1. Load staging table
+        loadBasicStagingData(dataPass1);
+        // 2. Execute plans and verify results
+        Map<String, Object> expectedStats = new HashMap<>();
+        expectedStats.put(StatisticName.INCOMING_RECORD_COUNT.name(), 3);
+        expectedStats.put(StatisticName.ROWS_DELETED.name(), 0);
+        expectedStats.put(StatisticName.ROWS_UPDATED.name(), 0);
+        expectedStats.put(StatisticName.ROWS_TERMINATED.name(), 0);
+        executePlansAndVerifyResults(ingestMode, options, datasets, schema, expectedDataPass1, expectedStats, fixedClock_2000_01_01);
+        // 3. Verify schema related changes
+        List<Map<String, Object>> actualTableData = h2Sink.executeQuery("select * from \"TEST\".\"main\"");
+        assertTableColumnsEquals(Arrays.asList(schema), actualTableData);
+    }
+
+    /*
+    @Test
+    void testAppendOnlyWithImplicitDataTypeChange() throws Exception
+    {
+        DatasetDefinition mainTable = TestUtils.getBasicMainTable();
+        DatasetDefinition stagingTable = TestUtils.getStagingTableForImplicitSchemaEvolution();
+
+        // Create staging table
+        createStagingTable(stagingTable);
+
+        // Generate the milestoning object
+        AppendOnly ingestMode = AppendOnly.builder()
+            .digestField(digestName)
+            .deduplicationStrategy(FilterDuplicates.builder().build())
+            .auditing(NoAuditing.builder().build())
+            .build();
+
+        PlannerOptions options = PlannerOptions.builder().cleanupStagingData(false).collectStatistics(true).enableSchemaEvolution(true).build();
+        Datasets datasets = Datasets.of(mainTable, stagingTable);
+
+        String[] schema = new String[]{idName, nameName, incomeName, startTimeName, expiryDateName, digestName};
+
+        // ------------ Perform incremental (append) milestoning Pass1 ------------------------
+        String dataPass1 = basePath + "input/vanilla_case/data_pass1.csv";
+        String expectedDataPass1 = basePath + "expected/vanilla_case/expected_pass1.csv";
+        // 1. Load staging table
+        loadBasicStagingData(dataPass1);
+        // 2. Execute plans and verify results
+        Map<String, Object> expectedStats = new HashMap<>();
+        expectedStats.put(StatisticName.INCOMING_RECORD_COUNT.name(), 3);
+        expectedStats.put(StatisticName.ROWS_DELETED.name(), 0);
+        expectedStats.put(StatisticName.ROWS_UPDATED.name(), 0);
+        expectedStats.put(StatisticName.ROWS_TERMINATED.name(), 0);
+
+        executePlansAndVerifyResults(ingestMode, options, datasets, schema, expectedDataPass1, expectedStats);
+        // 3. Assert that the staging table is NOT truncated
+        List<Map<String, Object>> stagingTableList = h2Sink.executeQuery("select * from \"TEST\".\"staging\"");
+        Assertions.assertEquals(stagingTableList.size(), 3);
+
+        // ------------ Perform incremental (append) milestoning Pass2 ------------------------
+        String dataPass2 = basePath + "input/vanilla_case/data_pass2.csv";
+        String expectedDataPass2 = basePath + "expected/vanilla_case/expected_pass2.csv";
+        // 1. Load staging table
+        loadBasicStagingData(dataPass2);
+        // 2. Execute plans and verify results
+        executePlansAndVerifyResults(ingestMode, options, datasets, schema, expectedDataPass2, expectedStats);
+    }
+
+     */
+}
