+++ legend-engine-xt-persistence-component/legend-engine-xt-persistence-component-relational-h2/src/test/resources/data/schema-evolution/input/add_data_pass.csv
@@ -1,0 +1,3 @@
+1,HARRY,1000,2020-01-01 00:00:00.0,2022-12-01,DIGEST1
+2,ROBERT,2000,2020-01-02 00:00:00.0,2022-12-02,DIGEST2
+3,ANDY,3000,2020-01-03 00:00:00.0,2022-12-03,DIGEST3
